requirements.txt
streamlit>=1.25
pandas>=1.5
numpy>=1.24
requests>=2.28
yfinance>=0.2.18
google-generativeai>=0.5.0
